Allow index.rst files in course webpages
Edit rst_creator and thumbnail_maker to allow index.rst files in
course webpages.  If a directory has an index.rst file, that file is
processed using rst2html and nothing else is done with that directory.
It is assumed that the presence of the index.rst file means you don't
want to automatically generate an html page for that directory.